# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DESIGN.Studio API — print-first AI design studio
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/palettes:
    get:
      operationId: listPalettes
      tags: [colors]
      summary: List color palettes for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Palettes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ColorPalette"
    post:
      operationId: createPalette
      tags: [colors]
      summary: Save a color palette
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ColorPaletteInput"
      responses:
        "201":
          description: Created palette
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColorPalette"
  /colors/extract:
    post:
      operationId: extractColors
      tags: [colors]
      summary: Extract dominant colors from an image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ColorExtractInput"
      responses:
        "200":
          description: Extracted colors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ColorExtractResult"

  # ── AI Jobs ────────────────────────────────────────────────────────────────
...
  /print-jobs:
    get:
      operationId: listPrintJobs
      tags: [printJobs]
      summary: List print jobs
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Print jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrintJob"
    post:
      operationId: createPrintJob
      tags: [printJobs]
      summary: Create a print separation job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrintJobInput"
      responses:
        "201":
          description: Created print job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintJob"
  /print-jobs/{id}:
    get:
      operationId: getPrintJob
      tags: [printJobs]
      summary: Get print job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Print job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrintJob"

  # ── Tech Pack ──────────────────────────────────────────────────────────────
...
  /manufacturing/pricing:
    get:
      operationId: getManufacturingPricing
      tags: [manufacturing]
      summary: Get pricing breakdown for a product
      parameters:
        - name: productType
          in: query
          schema:
            type: string
        - name: quantity
          in: query
          schema:
            type: integer
        - name: printMethod
          in: query
          schema:
            type: string
            enum: [screen_print, dtg, embroidery, sublimation]
      responses:
        "200":
          description: Pricing breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingBreakdown"
  /manufacturing/manufacturers:
    get:
      operationId: listManufacturers
      tags: [manufacturing]
      summary: List manufacturers directory
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [pod, screen_print, embroidery, mixed]
        - name: specialty
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Manufacturers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Manufacturer"
  /manufacturing/orders:
    get:
      operationId: listOrders
      tags: [manufacturing]
      summary: List manufacturing orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManufacturingOrder"
    post:
      operationId: createOrder
      tags: [manufacturing]
      summary: Place a manufacturing order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManufacturingOrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManufacturingOrder"

  # ── Collections ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ColorSwatch:
      type: object
      required: [hex, rgb, name]
      properties:
        hex:
          type: string
        rgb:
          type: string
        cmyk:
          type: ["string", "null"]
        pantone:
          type: ["string", "null"]
        name:
          type: string
        percentage:
          type: ["number", "null"]
    ColorPalette:
      type: object
      required: [id, projectId, name, colors, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        colors:
          type: array
          items:
            $ref: "#/components/schemas/ColorSwatch"
        createdAt:
          type: string
          format: date-time
    ColorPaletteInput:
      type: object
      required: [name, colors]
      properties:
        name:
          type: string
        colors:
          type: array
          items:
            $ref: "#/components/schemas/ColorSwatch"
    ColorExtractInput:
      type: object
      required: [assetUrl]
      properties:
        assetUrl:
          type: string
        colorCount:
          type: integer
          default: 6
    ColorExtractResult:
      type: object
      required: [colors]
      properties:
        colors:
          type: array
          items:
            $ref: "#/components/schemas/ColorSwatch"

    # ── AI Jobs ───────────────────────────────────────────────────────────────
...
    PrintJob:
      type: object
      required: [id, projectId, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        sourceAssetUrl:
          type: ["string", "null"]
        channelCount:
          type: integer
        lpi:
          type: integer
        dotShape:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, processing, completed, failed]
        outputZipUrl:
          type: ["string", "null"]
        channels:
          type: array
          items:
            $ref: "#/components/schemas/PrintChannel"
        createdAt:
          type: string
          format: date-time
    PrintChannel:
      type: object
      required: [name, color, angle]
      properties:
        name:
          type: string
        color:
          type: string
        angle:
          type: number
        outputUrl:
          type: ["string", "null"]
    PrintJobInput:
      type: object
      required: [projectId, sourceAssetUrl]
      properties:
        projectId:
          type: integer
        sourceAssetUrl:
          type: string
        channelCount:
          type: integer
          default: 4
        lpi:
          type: integer
          default: 65
        dotShape:
          type: string
          enum: [round, ellipse, line, diamond]
          default: round
        channels:
          type: array
          items:
            $ref: "#/components/schemas/PrintChannel"

    # ── Tech Pack ─────────────────────────────────────────────────────────────
...
    PricingBreakdown:
      type: object
      required: [baseCost, printingCost, shippingCost, totalCogs, suggestedRetail, marginPercent]
      properties:
        baseCost:
          type: number
        printingCost:
          type: number
        shippingCost:
          type: number
        platformFees:
          type: number
        totalCogs:
          type: number
        suggestedRetail:
          type: number
        marginPercent:
          type: number
        bulkTiers:
          type: array
          items:
            $ref: "#/components/schemas/BulkTier"
    BulkTier:
      type: object
      required: [quantity, pricePerUnit, totalCost]
      properties:
        quantity:
          type: integer
        pricePerUnit:
          type: number
        totalCost:
          type: number
        marginAtRetail:
          type: ["number", "null"]
    Manufacturer:
      type: object
      required: [id, name, type, turnaround]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [pod, screen_print, embroidery, mixed]
        website:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        moq:
          type: ["integer", "null"]
        turnaround:
          type: string
        sustainable:
          type: boolean
        specialties:
          type: array
          items:
            type: string
        countries:
          type: array
          items:
            type: string
        hasApi:
          type: boolean
        rating:
          type: ["number", "null"]
    ManufacturingOrder:
      type: object
      required: [id, projectId, manufacturerId, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        manufacturerId:
          type: integer
        status:
          type: string
          enum: [draft, submitted, confirmed, in_production, shipped, delivered, cancelled]
        quantity:
          type: integer
        totalCost:
          type: ["number", "null"]
        externalOrderId:
          type: ["string", "null"]
        trackingUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ManufacturingOrderInput:
      type: object
      required: [projectId, manufacturerId, quantity]
      properties:
        projectId:
          type: integer
        manufacturerId:
          type: integer
        quantity:
          type: integer
        notes:
          type: string
        mockupAssetUrl:
          type: string

    # ── Collections ───────────────────────────────────────────────────────────
...