# lib/api-zod/src/generated/types/colorExtractResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DESIGN.Studio API — print-first AI design studio
 * OpenAPI spec version: 0.1.0
 */
import type { ColorSwatch } from './colorSwatch';

export interface ColorExtractResult {
  colors: ColorSwatch[];
}