# lib/api-zod/src/generated/types/colorSwatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DESIGN.Studio API — print-first AI design studio
 * OpenAPI spec version: 0.1.0
 */

export interface ColorSwatch {
  hex: string;
  rgb: string;
  /** @nullable */
  cmyk?: string | null;
  /** @nullable */
  pantone?: string | null;
  name: string;
  /** @nullable */
  percentage?: number | null;
}